+++ CHANGELOG.md
@@ -21,6 +21,16 @@ See [CHANGELOG - v7](https://github.com/webdriverio/webdriverio/blob/v7/CHANGELO
 
 ---
 
+## v8.29.0 (2024-01-22)
+
+#### :rocket: New Feature
+* `wdio-browser-runner`, `wdio-cli`, `wdio-globals`, `wdio-jasmine-framework`, `wdio-runner`, `wdio-types`
+  * [#12058](https://github.com/webdriverio/webdriverio/pull/12058) (@wdio/cli): add support for snapshot testing ([@christian-bromann](https://github.com/christian-bromann))
+
+#### Committers: 1
+- Christian Bromann ([@christian-bromann](https://github.com/christian-bromann))
+
+
 ## v8.28.8 (2024-01-20)
 
 #### :bug: Bug Fix